requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
albumentations>=1.3.0
timm>=0.9.0
reservoirpy>=0.3.0
imbalanced-learn>=0.11.0
joblib>=1.3.0
flask>=2.3.0
reportlab>=4.0.0
kaggle>=1.5.0
pyyaml>=6.0
tqdm>=4.65.0
